Refactored algorithms module into a sub-package, and split all algorithms in dedicated modules.
The API has also changed, which will require a major release bump
