+++ examples/01_camera.py
@@ -47,22 +47,24 @@ print(f"Img shape: {img_orig.shape}")
 
 net_params = ad.NetworkParamsUNet(n_features=16)
 
-denoiser_sup = ad.Denoiser(model=net_params, reg_val=REG_TV_VAL)
-denoiser_sup.train_supervised(imgs_noisy, img_orig, epochs=EPOCHS, tst_inds=tst_inds)
-
-denoiser_n2n = ad.N2N(model=net_params, reg_val=REG_TV_VAL)
-denoiser_n2n.train_selfsupervised(imgs_noisy, epochs=EPOCHS)
+denoiser_sup = ad.Supervised(model=net_params, reg_val=REG_TV_VAL)
+denoiser_sup.train(imgs_noisy, img_orig, epochs=EPOCHS, tst_inds=tst_inds)
 
 denoiser_n2v = ad.N2V(model=net_params, reg_val=REG_TV_VAL)
-denoiser_n2v.train_selfsupervised(imgs_noisy, epochs=EPOCHS, tst_inds=tst_inds)
+denoiser_n2v.train(imgs_noisy, epochs=EPOCHS, tst_inds=tst_inds)
+
+denoiser_n2n = ad.N2N(model=net_params, reg_val=REG_TV_VAL)
+n2n_data = denoiser_n2n.prepare_input(imgs_noisy)
+denoiser_n2n.train(*n2n_data, epochs=EPOCHS)
 
 denoiser_dip = ad.DIP(model=net_params, reg_val=REG_TV_VAL)
-inp_dip = denoiser_dip.train_unsupervised(imgs_noisy, epochs=EPOCHS)
+dip_data = denoiser_dip.prepare_input(imgs_noisy)
+denoiser_dip.train(*dip_data, epochs=EPOCHS)
 
 den_sup = denoiser_sup.infer(imgs_noisy).mean(0)
-den_n2n = denoiser_n2n.infer(imgs_noisy).mean(0)
 den_n2v = denoiser_n2v.infer(imgs_noisy).mean(0)
-den_dip = denoiser_dip.infer(inp_dip)
+den_n2n = denoiser_n2n.infer(n2n_data[0]).mean(0)
+den_dip = denoiser_dip.infer(dip_data[0])
 
 fig, axs = plt.subplots(2, 3, sharex=True, sharey=True)
 axs[0, 0].imshow(img_orig)
@@ -71,10 +73,10 @@ axs[0, 1].imshow(imgs_noisy[0])
 axs[0, 1].set_title("Noisy image")
 axs[0, 2].imshow(den_sup)
 axs[0, 2].set_title("Denoised supervised")
-axs[1, 0].imshow(den_n2n)
-axs[1, 0].set_title("Denoised N2N")
-axs[1, 1].imshow(den_n2v)
-axs[1, 1].set_title("Denoised N2V")
+axs[1, 0].imshow(den_n2v)
+axs[1, 0].set_title("Denoised N2V")
+axs[1, 1].imshow(den_n2n)
+axs[1, 1].set_title("Denoised N2N")
 axs[1, 2].imshow(den_dip)
 axs[1, 2].set_title("Denoised DIP")
 fig.tight_layout()
